Use mod in place of use
`use` is equivalent to `using` in C# or `using namespace` in C++, in
that it's pretty much used to shorten paths. `mod X`, on the other hand,
looks for a file named X (or X/mod.rs) and inserts its contents as a new
module in the current file. It seems like `mod commands` is more
appropriate for this use case, and the more qualified name of
`commands::build_framework` helps avoid ambiguity.